packages/ocr_utils/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "ocr_utils"
dynamic = ["version"]
description = "Shared OCR processing utilities using OCR service or macOS LiveText fallback"
readme = "README.md"
authors = [
    {name = "Jeffrey Urban"}
]
requires-python = ">=3.13,<3.14"
keywords = [
    "ocr",
    "text-recognition",
    "google-cloud-vision"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Graphics :: Capture",
    "Topic :: Text Processing",
    "Topic :: Utilities",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14"
]
dependencies = [
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.27.0"
]

[project.scripts]
ocr_utils = "ocr_utils.cli:app"

[project.optional-dependencies]
ocrmac = [
    "ocrmac>=0.1.0"
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.9",
    "pyright>=1.1.0"
]

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
root = "../.."

[tool.hatch.build.targets.wheel]
packages = ["src/ocr_utils"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests requiring macOS LiveText"
]

[tool.coverage.run]
source = ["src/ocr_utils"]
omit = ["*/tests/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:"
]

[tool.hatch.build.hooks.vcs]
version-file = "src/ocr_utils/_version.py"
